update the deployment script to scan secrets

+++ scripts/app_deployment.sh
@@ -9,8 +9,18 @@ set -e
 
 source avenv/bin/activate
 
+secret_scan_results=$(detect-secrets scan | \
+python3 -c "import sys, json; print(json.load(sys.stdin)['results'])" )
+
+# static scan for security credentials that terminates if any secrets are found
+if [ "${secret_scan_results}" != "{}" ]; then
+    echo "detect-secrets scan failed"
+    exit 125
+fi
+
 python -m unittest
 
+# removes the deployment .zip package locally if it exists
 if [ -e $DEPLOYMENT_PACKAGE ]; then
     rm $DEPLOYMENT_PACKAGE
 fi
